Rename csproj to not clash with endpoint class
