+++ samples/azure-service-bus-netstandard/lock-renewal/ASBS_1/LockRenewal/LockRenewal.csproj
@@ -3,6 +3,7 @@
     <TargetFrameworks>netcoreapp3.1;netcoreapp2.1;net48</TargetFrameworks>
     <OutputType>Exe</OutputType>
     <LangVersion>7.3</LangVersion>
+    <RootNamespace>Endpoint</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="NServiceBus.Transport.AzureServiceBus" Version="1.*" />
