+++ samples/azure-service-bus-netstandard/lock-renewal/ASBS_1/LockRenewal/LockRenewalConfiguration.cs
