+++ samples/azure-service-bus-netstandard/lock-renewal/ASBS_1/LockRenewal/LockRenewalOptions.cs
