+++ samples/azure-service-bus-netstandard/lock-renewal/ASBS_1/LockRenewal/Program.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
-using Endpoint;
+using LockRenewal;
 using Microsoft.Azure.ServiceBus.Management;
 using NServiceBus;
 
@@ -34,7 +34,7 @@ class Program
 
         #endregion
 
-        var endpointInstance = await NServiceBus.Endpoint.Start(endpointConfiguration).ConfigureAwait(false);
+        var endpointInstance = await Endpoint.Start(endpointConfiguration).ConfigureAwait(false);
 
         await OverrideQueueLockDuration("Samples.ASB.SendReply.LockRenewal", connectionString).ConfigureAwait(false);
