+++ samples/azure-service-bus-netstandard/lock-renewal/ASBS_2/LockRenewal/LockRenewalBehavior.cs
