+++ samples/azure-service-bus-netstandard/lock-renewal/ASBS_2/LockRenewal/LockRenewalOptions.cs
